# packages/client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Corpact API",
    "version": "0.1.0",
    "description": "Corporate-action accounting for tokenized equities on Solana. Quantities, raw amounts, slots and USD values are decimal strings; null means unknown, never zero."
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Corpact API key (cpk_…)"
      }
    },
...
  },
  "paths": {
...
    "/v1/assets": {
      "get": {
        "summary": "Supported mints and their verification status",
        "tags": [
          "assets"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "assets"
                  ],
                  "properties": {
                    "assets": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "mint",
                          "symbol",
                          "name",
                          "decimals",
                          "registrySource",
                          "verified",
                          "verificationError",
                          "verifiedSlot",
                          "capabilities"
                        ],
                        "properties": {
                          "mint": {
                            "type": "string"
                          },
                          "symbol": {
                            "type": "string"
                          },
                          "name": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "decimals": {
                            "type": [
                              "integer",
                              "null"
                            ]
                          },
                          "registrySource": {
                            "type": "string",
                            "enum": [
                              "fixtures",
                              "live"
                            ]
                          },
                          "verified": {
                            "type": "boolean"
                          },
                          "verificationError": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "verifiedSlot": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "pattern": "^-?\\d+(\\.\\d+)?$"
                          },
                          "capabilities": {
                            "type": "object",
                            "additionalProperties": false,
                            "required": [
                              "tracking",
                              "manualConversion",
                              "automation"
                            ],
                            "properties": {
                              "tracking": {
                                "type": "boolean"
                              },
                              "manualConversion": {
                                "type": "boolean"
                              },
                              "automation": {
                                "type": "boolean"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "missing_scope",
                        "wallet_not_registered",
                        "wallet_quota_exceeded"
                      ],
                      "description": "Machine-readable reason, when a client should branch on it"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "missing_scope",
                        "wallet_not_registered",
                        "wallet_quota_exceeded"
                      ],
                      "description": "Machine-readable reason, when a client should branch on it"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "missing_scope",
                        "wallet_not_registered",
                        "wallet_quota_exceeded"
                      ],
                      "description": "Machine-readable reason, when a client should branch on it"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "missing_scope",
                        "wallet_not_registered",
                        "wallet_quota_exceeded"
                      ],
                      "description": "Machine-readable reason, when a client should branch on it"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "missing_scope",
                        "wallet_not_registered",
                        "wallet_quota_exceeded"
                      ],
                      "description": "Machine-readable reason, when a client should branch on it"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    }
  ]
}